requirements.txt
asgiref==3.7.2
certifi==2023.7.22
chardet==3.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-filter==23.2
django-import-export==3.2.0
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
environs==9.5.0
et-xmlfile==1.1.0
Faker==19.6.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
MarkupPy==1.14
marshmallow==3.20.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
